pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kalmanorix"
version = "0.1.0"
description = "Kalmanorix scaffold"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["numpy>=1.26"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pylint>=3.0", "ruff>=0.6", "black>=24.0", "mypy>=1.8", "pre-commit>=3.6", "types-PyYAML>=6.0"]
viz = ["matplotlib>=3.7"]
train = [
    "sentence-transformers>=3.0",
    "torch>=2.0",
    "datasets>=2.14",
    "transformers>=4.36",
    "accelerate>=0.25",
    "pyyaml>=6.0",
    "psutil>=5.9"
]
cloud = [
    "openai>=1.0",
    "cohere>=5.0",
    "anthropic>=0.25",
    "google-cloud-aiplatform>=1.50"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
